dump: Update polymer to fix chrome trace html files
The output of trace2html uses polymer inside, but some of the features
are disabled in the recent Chrome browsers, maybe Chrome 80.

It shows the following error.

  Uncaught TypeError: document.registerElement is not a function

The document.registerElement is deprecated in the recent Chrome browser.

This patch updates the broken parts to the newer html contents same as
deno.html and it fixes all the problems.

Signed-off-by: Honggyu Kim <honggyu.kp@gmail.com>